Math Quiz - Audio + Score
Added some "correct answer" audio. It's also possible to add a "wrong answer" audio file, but need to find the right one yet. Correct score and incorrect score are now shown on the left and right side of the question within the panel. Also fixed a bug where there was a spacebar too much within the questions.